invoice_comparison/requirements.txt
Flask==2.1.0
pytesseract==0.3.8
pdf2image==1.16.0
Pillow==8.4.0
